LUN-2829 Handle workflows that do not create the attribute unsaved_select_profiles.
This attribute is set with the profiles that are selected by the user.
The attribute is necessary because the many-to-one relationships cannot be created before the profile promo grid is not saved, so we cannot do it in the form save.
When the plugin is copied, this attribute will be missing, so that is the reason of the hasattr.